+++ dgisim/tests/test_card/test_instructors_cap.py
@@ -1,4 +1,3 @@
-
 import unittest
 
 from .common_imports import *