+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Removes preceding `v` in version name to be like `0.3.2`.
       - Wolf's Gravestone
     - Artifact Card:
       - Instructor's Cap
+      - Tenacity of the Millelith
   - Event Card:
     - Where Is the Unseen Razor?